fix .babelrc issue with rollup + mocha
Interestingly, the use of rollup dictates that “module” support in babel be disabled - this is terrible news for mocha, which sort of appreciates them.

I’ve instead moved the babel config to the rollup script (with the additional request that it doesn’t use any other .babelrc files) and then let mocha use the root level .babelrc as usual.